refactor(product, promotion): Product와 Promotion의 공통 읽기 로직 추출
Product와 Promotion 간의 코드 중복을 줄이기 위해 Stream을 통한 공통 읽기 기능을 별도의 FileUtils 유틸리티로 분리했습니다.
