+++ README.md
@@ -7,7 +7,7 @@
 ### Product
 
 - 상품의 수량이 0 미만이면 안된다.
-- 재고수량이 달라도 상품명이 같으면 같은 상품이다.
+- 상품명과 프로모션이 같으면 같은 상품이다.
 - 같은 상품이 존재해서는 안된다.
 
 ### ProductParser
