+++ src/main/java/domain/utils/FileUtils.java
@@ -1,0 +1,19 @@
+package domain.utils;
+
+import java.io.File;
+import java.io.FileNotFoundException;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Scanner;
+
+public class FileUtils {
+    public static List<String> readLinesFromFile(final String filePath) throws FileNotFoundException {
+        List<String> lines = new ArrayList<>();
+        try (Scanner scanner = new Scanner(new File(filePath))) {
+            while (scanner.hasNextLine()) {
+                lines.add(scanner.nextLine());
+            }
+        }
+        return lines;
+    }
+}
